+++ coloraide/css/parse.py
@@ -4,7 +4,7 @@ import math
 from .. import algebra as alg
 from ..types import Vector
 from . import color_names
-from ..gamut.bounds import FLG_ANGLE, FLG_PERCENT, FLG_OPT_PERCENT
+from ..channels import FLG_ANGLE, FLG_PERCENT, FLG_OPT_PERCENT
 from typing import Optional, Tuple
 from typing import Dict, Type, TYPE_CHECKING
 
@@ -344,7 +344,10 @@ def parse_css(
                     return (value[:3], value[3]), m.end(0)
             else:
                 offset = m.start(0)
-                return parse_rgb_channels(string[m.end(1) - offset + 1:m.end(0) - offset - 1], cspace.CHANNELS), m.end(0)
+                return (
+                    parse_rgb_channels(string[m.end(1) - offset + 1:m.end(0) - offset - 1], cspace.CHANNELS),
+                    m.end(0)
+                )
     else:
         m = CSS_MATCH[cspace.NAME].match(string, start)
         if m is not None and (not fullmatch or m.end(0) == len(string)):
