+++ coloraide/spaces/lch/__init__.py
@@ -1,7 +1,7 @@
 """Lch class."""
-from ...spaces import Space, Lchish, Channel, chroma_coord
+from ...spaces import Space, Lchish
 from ...cat import WHITES
-from ...gamut.bounds import GamutUnbound, FLG_ANGLE, FLG_OPT_PERCENT
+from ...channels import Channel, FLG_ANGLE, FLG_OPT_PERCENT
 from ... import util
 import math
 from ... import algebra as alg
