+++ coloraide/spaces/oklab/__init__.py
@@ -25,9 +25,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from ...spaces import Space, Labish, Channel
+from ...spaces import Space, Labish
 from ...cat import WHITES
-from ...gamut.bounds import GamutUnbound, FLG_OPT_PERCENT
+from ...channels import Channel, FLG_OPT_PERCENT
 from ... import algebra as alg
 from ...types import Vector
