+++ coloraide/spaces/oklch/__init__.py
@@ -23,9 +23,9 @@ LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
 SOFTWARE.
 """
-from ...spaces import Space, Lchish, Channel, chroma_coord
+from ...spaces import Space, Lchish
 from ...cat import WHITES
-from ...gamut.bounds import GamutUnbound, FLG_ANGLE, FLG_OPT_PERCENT
+from ...channels import Channel, FLG_ANGLE, FLG_OPT_PERCENT
 from ... import util
 import math
 from ... import algebra as alg
