+++ tools/interp_diagram.py
@@ -103,7 +103,7 @@ def main():
     name1 = args.xaxis.split(":", 1)[0]
     name2 = args.yaxis.split(":", 1)[0]
     name1 = c1._space.CHANNEL_ALIASES.get(name1, name1)
-    index1 = c1._space.CHANNEL_NAMES.index(name1)
+    index1 = c1._space.get_channel_index(name1)
     name2 = c1._space.CHANNEL_ALIASES.get(name2, name2)
     hue_index = -1
